Combining left and right base rules.
