+++ vsg/rules/__init__.py
@@ -21,6 +21,7 @@ from .blank_line_below_line_ending_with_several_possible_tokens import blank_lin
 from .insert_token_left_of_token_if_it_does_not_exist_between_tokens_using_value_from_token import insert_token_left_of_token_if_it_does_not_exist_between_tokens_using_value_from_token
 from .insert_token_left_of_token_if_it_does_not_exist_between_tokens import insert_token_left_of_token_if_it_does_not_exist_between_tokens
 from .insert_token_right_of_token_if_it_does_not_exist_between_tokens_using_value_from_token import insert_token_right_of_token_if_it_does_not_exist_between_tokens_using_value_from_token
+from .insert_token_next_to_token_if_it_does_not_exist_between_tokens_using_value_from_token import insert_token_next_to_token_if_it_does_not_exist_between_tokens_using_value_from_token
 from .is_token_value_one_of import is_token_value_one_of
 from .align_tokens_in_region_between_tokens import align_tokens_in_region_between_tokens
 from .align_tokens_in_region_between_tokens_unless_between_tokens import align_tokens_in_region_between_tokens_unless_between_tokens
