+++ vsg/rules/insert_token_right_of_token_if_it_does_not_exist_between_tokens_using_value_from_token.py
@@ -1,5 +1,4 @@
 
-
 from vsg import parser
 from vsg import violation
 
@@ -7,8 +6,10 @@ from vsg.vhdlFile import utils
 from vsg.rule_group import structure
 from vsg.rules import utils as rules_utils
 
+from vsg.rules.insert_token_next_to_token_if_it_does_not_exist_between_tokens_using_value_from_token import insert_token_next_to_token_if_it_does_not_exist_between_tokens_using_value_from_token as Rule
 
-class insert_token_right_of_token_if_it_does_not_exist_between_tokens_using_value_from_token(structure.Rule):
+
+class insert_token_right_of_token_if_it_does_not_exist_between_tokens_using_value_from_token(Rule):
     '''
     Checks for the existence of a token and will insert it if it does not exist.
 
@@ -32,93 +33,11 @@ class insert_token_right_of_token_if_it_does_not_exist_between_tokens_using_valu
     '''
 
     def __init__(self, name, identifier, insert_token, anchor_token, left_token, right_token, value_token):
-        structure.Rule.__init__(self, name=name, identifier=identifier)
+        Rule.__init__(self, name=name, identifier=identifier)
         self.insert_token = insert_token
         self.anchor_token = anchor_token
         self.left_token = left_token
         self.right_token = right_token
         self.value_token = value_token
-        self.action = 'add'
-        self.configuration.append('action')
         self.configuration_documentation_link = 'configuring_optional_items_link'
-        self.filter_tokens = []
-
-    def _get_tokens_of_interest(self, oFile):
-        if remove_keyword(self):
-            return oFile.get_token_and_n_tokens_before_it([self.insert_token], 1)
-        else:
-            lToi = oFile.get_tokens_between_tokens_inclusive_while_storing_value_from_token(self.left_token, self.right_token, self.value_token)
-            lReturn = []
-            for oToi in lToi:
-                for oToken in self.filter_tokens:
-                    if oToi.token_type_exists(oToken):
-                        break
-                else:
-                    lReturn.append(oToi)
-            return lReturn
-
-    def _analyze(self, lToi):
-        if remove_keyword(self):
-            analyze_for_existence_of_optional_keyword(lToi, self)
-        else:
-            analyze_for_missing_optional_keyword(lToi, self)
-
-    def _fix_violation(self, oViolation):
-        if remove_keyword(self):
-            rules_utils.remove_optional_item(oViolation, self.insert_token)
-        else:
-            add_optional_item(oViolation, self)
-
-
-def remove_keyword(self):
-    if self.action == 'remove':
-        return True
-    return False
-
-
-def analyze_for_existence_of_optional_keyword(lToi, self):
-    for oToi in lToi:
-        oViolation = create_violation(oToi, oToi.get_line_number(), self)
-        self.add_violation(oViolation)
-
-
-def analyze_for_missing_optional_keyword(lToi, self):
-    for oToi in lToi:
-        iLine, lTokens = utils.get_toi_parameters(oToi)
-        if not optional_keyword_exists(self.insert_token, lTokens):
-            iLine += rules_utils.get_number_of_carriage_returns_before_token(self.anchor_token, lTokens)
-            oViolation = create_violation(oToi, iLine, self)
-            self.add_violation(oViolation)
-
-
-def optional_keyword_exists(oToken, lTokens):
-    if rules_utils.get_index_of_token_in_list(oToken, lTokens) is None:
-        return False
-    return True
-
-
-def add_optional_item(oViolation, self):
-    lTokens = oViolation.get_tokens()
-    if not token_value_available(oViolation):
-        return
-
-    if not optional_keyword_exists(self.anchor_token, lTokens):
-        return
-
-    iIndex = rules_utils.get_index_of_token_in_list(self.anchor_token, lTokens)
-
-    rules_utils.insert_token(lTokens, iIndex + 1, self.insert_token(oViolation.get_token_value()))
-    rules_utils.insert_whitespace(lTokens, iIndex + 1)
-    oViolation.set_tokens(lTokens)
-
-
-def token_value_available(oViolation):
-    if oViolation.get_token_value() is not None:
-        return True
-    return False
-
-
-def create_violation(oToi, iLineNumber, self):
-    sSolution = self.action.capitalize() + ' ' + self.solution
-    oViolation = violation.New(iLineNumber, oToi, sSolution)
-    return oViolation
+        self.direction = 'right'
